Change directory names of testing suite
The directory 'test_cases/ocean' was originally a temporary name while
the ocean tested these scripts.  It is renamed here:

test_cases/ocean -> testing_and_setup/compass

The name indicates that both testing and setup are done with these
scripts.  COMPASS is a new acronym for
"Configuration Of Model for Prediction Across Scales Setups"
and is one of several potential tools that could be placed in the new
testing_and_setup directory.