Set temperature to constant under land ice
This merge adds a constant to global_ocean init that sets the
temperature under ice shelves to a constant value (close to
freezing).  This feature has been turned on in cases with land-ice
cavities in the test-cases infrastructure.

Surface restoring has also been turned off under land ice because
it has undesired consequences.