Add 1-8km configuration for Thwaites sims
This commit adds the creation of a 1-8km variable resolution mesh for
Thwaites Glacier using the JIGSAW mesh generation tool.  JIGSAW needs to
be run manually first, but then the test case infrastructure is set up
to automate the rest of the processing steps to finalize the mesh and
the run configuration.