Adjust finalize step for Thwaites mesh to remove embayments
With the new 1km mesh, there were some weird embayments into the ice
shelf that were only 1 cell wide.  For some reason, these location
experience unrealistically large velocities, about twice as fast as the
surrounding shelf.  The model ran fine, but the time step had to be
quite small to accommodate this.  The high velocity probably has
something to do with how we extrapolate the state from cells with ice
forward for Albany, combined with the fact that we have a 'kill' value
in the SMB outside the initial ice extent, which maintains these
embayments.

This commit adds a step to the finalize script for the thwaites mesh
that identifies these narrow 1-cell wide embayments and fills them in
with ice in the initial condition file.  This modification has a
negliblble impact on the overall ice geometry and it eliminates the
fast velocities.  To the necessary calculations, the finalize script has
been converted from a bash script using NCO to a python script.