Add var res dome mesh script for Jigsaw
This Matlab script was used to create a variable resolution landice grid
for the Halfar test case.  Results from this grid are compared to
uniform resolution meshes in the MPASLI description paper.  The grid
creation steps after JIGSAW were done manually, but could be automated.

The density function is specified as a radial distance from the center
of the dome.