Update HO decomp test criteria
The HO decomposition tests are not BFB due to problems in Albany.  Due
to changes in Albany, the tests have become even more non-BFB.  Rather
than have failing test, I am decreasing the tolerance of these tests for
this known problem.  That way, the test results are still useful as
regression tests.