Point global_ocean_V1 test cases to correct templates
Many test cases are trying to use templates from 'global_ocean',
which works for critical passages (but is not necessary, since
'global_ocean_V1' has a copy of the same template, and could lead
to issues in the future) but not for init2, since 'global_ocean'
does not use this template.

A few test cases refer to the invalid "compass" base path.