Switch precond. from ML to Ifpack for restart tests
The ML preconditioner in Albany currently is unable to be
bit-restartable due to the usage of a random number generator to
initialize the search for eigenvalues.  This will eventually be fixed to
allow a specified seed value.

Until then, switch from ML to Ifpack to allow for bit-restartability
testing to work.