Rename waterfrac to waterFrac and make thermal package
This commit renames waterfrac to waterFrac everywhere in the code.

It also adds a thermal package that prevents thermal-related variables
from being in restart files if the thermal solver is off.  Note that
temperature is still always included because it can be used to calculate
the flow rate factor even when temperature evolution is off.

This commit also adds a restart test for EISMINT2 (which currently
fails).