Update scripts to remove some subprocess calls
The paraview extractor can now be called as a function rather
than a script, and this is done during base-mesh generation
and culling.

SCRIP files can now also be created with a function, so a script
call is replaced with a function here as well.

With these changes, calls to python scripts that use NetCDF in
the parallel conda enviornment will now work as long as they
are called with `mpirun -np 1`