Switch E3SM coupling to mesh file
Most steps in E3SM need a mesh file, which can always come from
the end of `initial_state`.

Where an initial condition is needed instead of a mesh file, this
is taken from the config file.  Currently, the default is the same
as the mesh file: `../initial_state/initial_state.nc` but this
could be changed by the user (or a case-specific config file) in
the future.