Add "examples" core
These examples include a compact and an expanded example
configuration, each with two testcases with two steps.  One shows
how a testcase and its steps could be defined in a fully "expanded"
format with quite a bit of redundant code.  The other shows how to
write the same tests in a more compact, if perhaps less intuitive,
way.