Change the way steps run on their own
The key ``steps_to_run`` has been added to the ``testcase`` dict
and individual steps run the same way as the whole testcase except
that there is only one step in ``steps_to_run`` and logging is
handled slightly differently.

This means that config options related to the requested number of
cores, minimum cores, threads, etc. can be parsed at the testcase
level, then the number of cores to run with can be determined
in ``run_steps`` before running individual steps.

Testcases that run all steps in the order they are added can just
leave ``steps_to_run`` with its default value.  If there are some
optional steps that should be excluded, the testcase will need
to replace the default list with an appropriate list of steps.