Rewrite run_model to take the step dict and do more
Since we almost always want to:
* update PIO
* make a graph partiton
* run the model
these first two are now an optional part of running the model.
By having run_model take `step`, we can get information like the
number of cores, threads and work dir. from there.

To make this easier, the core and configration were added to `step`
while setting up `testcase` defaults.  There could be a little
confusion between "core" and "cores" (and "configuration" and
"config file" for that matter) but this is hard to avoid.