Make subdirs the keys for test case dicts
The name of the test case isn't always unique so it is necessary
to use a unique key like subdir (or path)

With this merge, only the values of the test_cases and steps dicts
are used in the framework but the dict datastructure ensures that
each test case has a unique subdirectory

Steps remain indexed by name becasue this is useful for finding
a specific step in a test case by name.  This requires that each
step have a unique name (for now, the same as the subdir).