Update User's Guide and Test Groups
This merge includes a large number of updates to the User's Guide
following recent moves to classes and a new naming convention.
Machine usage has also been updated to be consistent with recent
testing with E3SM modules.

The Developer's Guide has also been updated with the change
from "configuration" to "test group".