`compass run` runs a single suite without arg
With this merge, if you only set up a single test suite in a work
directory, `compass run` will work without the need to specify the
suite name.  Since this is by far the most common work flow, this
change should save a little trouble nearly every time a test suite
is run.