Simplify and generalize the download function
We don't use the ability to use config options for downloads
directly with this function, so this merge eliminates the unused
feature.

Cases have arisen where it is useful to have a URL that has a
different file name than the target filename that we want to
download to.  Sometimes, the URL doesn't have a particularly
unique or descriptive name.