Make HO velo versions of all dome tests
This commit makes versions of all dome tests that use the 'FO' velocity
solver instead of the 'sia' solver.  This generates tests that existed
in the old COMPASS system.  With the new system, it was possible to
reuse the existing dome cases, but simply add a new argumetn
'velo_solver' that indicates which velocity solver namelist option to
use.

This commit makes the required function changes, but the updating of the
namelist is failing.  That will be fixed in a subsequent commmit.  With
this commit, 'compass list' reports the new FO versions of these tests.