Adjust naming conventions and dun duration for dome tests
This commit modifies the dome test to follow a working dir naming
convention like: landice/dome/2000m/fo_restart_test
where one of ‘fo’ and ‘sia’ is used in the test name.  The test suites
are updated accordingly.
It also sets the default run duration to be 2 years, and specifically
modifies that to 200 years only for the sia smoke test.