Set up different validation norm values for different dome configurations
The dome decomposition tests using the FO solver are currently not BFB.
This commit modifies the validation step for the this test to support
per-configuration norm values.

Note I had to call the validation routine once for each variable being
tested to allow different norm values to be used for each variable.