Correct validation parsing of runs and simplify validation logic
This is based on the PR discussion with Xylar.  From that discussion:
"If you run just the 1proc_run step (in the SIA configuration) on its
own, which is a valid thing to do, you will trigger this error. So the
else should just be removed here. We only do validation if both steps
got run but it is valid to run only one of the steps."

This commit also simplifies the logic for validating the different velo
solvers (based on the PR discussion).