Add premade Thwaites 4km test group
This is a simple configuration that downloads a pre-existing 4km mesh
with an initial condition and then runs basic decomp and restart tests.
The tests are also added to the test suites.  This test only applies to
the FO velocity solver, so the velocity solver is not included in the
name or as a property of the test case.