Update the init test case
Store the `inactive_top_comp_subdir` so we know what files to
compare with during validation.

Use `ds.load()` when modifying `minLevelCell` in the initial state
to prevent issues with reading from and writing to the same file.

Crop inactive top cells after writing out the modified initial
state (because we now use the modified `minLevelCell`).

Set the default to `inactive_top_cells = 0` in a new `init.cfg`
instead of the config file for the QU240 mesh.