Remove critical_passages stream from initial_state
We do not deepen critical passages in any mesh configuration,
so the critical_passages mask is not needed during init. Requiring
the file in `stream.ocean` makes it difficult to swap the mesh
used in `initial_state` for another mesh (e.g. one generated on
a different machine or with a previous version of compass).