Print but don't raise errors in validation
When errors occur during variable validation such as a mis-match
in the number of cells, print an error and mark the test as failing
but don't raise an exception.  This way, the errors will appear
in the log and the test will be marked as failing but there will
not be a stack trace, nor will the suite output show that the
test execution (rather than baseline validation) failed.