Add `add_config` method to test cases
This new method is used to add a config file during test case
init in the same way that input, output, namelist and streams files
can be added to steps during their init.

This is useful because some config options should be computed in
the `configure()` method based on other config options, and it
may be helpful if (other than these computed config options) all
config options are loaded from files before `configure()` is
called.

Additionally, there is a list of dictionaries that keeps track of
the config files that are combined to get the config options for
a test case.  This may be useful in the future, e.g. for provenance.

Test-case setup has been modified to first call the `add_config`
method to build up a list of config files before loading them
more or less as usual.  The main difference is that all config
files are now loaded before calls to `configure()`, with the user
config file being added last.  This makes it more likely that
config options are as close as possible to their final state
before `configure()` is called.