Add support for compass test release
Move the config options related to depolyment of the conda env.
to the conda directory.

Make sure the shared environment gets updated and mamba/boa
installed.

If `env_type == 'test_release'` and recreating the environment,
also build the compass conda package and point to the path where
the package is built when creating the environment.