Switch to using symlink
This changes the previous commit to use the standalone database.  I
added a symlink on Anvil to the E3SM file for this purpose.  As Xylar
pointed out: "Otherwise, every time the test case gets set up, the mesh
will have to get downloaded again. If you add it to the database, it
will get cached locally."