Remove unnecessary code in mesh.py
Remove call to define_cullMask.py, as testing indicates
this step is not necessary. Also remove optional step to
subset data. This was not used and was potentially dangerous
if accidentally changed. Because it is easy enough to figure out
how to add back if needed, it is removed for now. Also fix a few
typos in Humboldt documentation.