Create decomposition tests for Humboldt, including calving
This commit adds a new composition test for the humboldt test group and
applies it to a range of calving laws.  The structure is based on the
dome decomposition test with the added argument for calving law.

This commit adds this new decomposition test to a combination of
different calving laws.  The code runs but there are a number of
details that need cleaning up in subsequent commits.