Update core counts for runs with particles
Particles requre a lot more memory to run successfully.  For now,
this is handled implicitly via the number of cores.  The minimum
values were determined for 5km and 2.5km resolution on Anvil by
trial and error.  The target values are then the sum of the
minimum with particles and the target minus the minimum without
particles.  Minimum cores for coarser resolution were determined
by dividing by 4 for each factor of 2 decrease in resolution.