Create set_cell_width() function
Create a function to determine cell_width based on input parameters in
the config file. This currently contains three different cell spacing
functions based on log10(speed), distance to ice edge, and distance to
the grounding line that can be turned on or off in the config file. Minimum
and maximum cell width and the distances over which cell width is varied
are also set in the config file. Implement a use case for Humboldt and update
User's Guide.