Add CompassConfigParser
This is a "meta" config parser that keeps a dictionary of
config parsers and their sources to combine when needed.  The
custom config parser allows provenance of the source of different
config options and allows the "user" config options to always
take precedence over other config options (even if they are
added later).