Add steps for making spherical meshes with JIGSAW
There is a base class and classes for creating either icosahedral
subdivision meshes or meshes from a map of cell width.

The `SphericalBaseStep` class contain an JIGSAW options
datastructure that specific implementations can modify to
configure how jigsaw will build the mesh.

The `QuasiUniformSphericalMeshStep` class creates a quasi-uniform
base mesh, and allows the `build_cell_width_lat_lon()` an
`make_jigsaw_mesh()` methods to be overridden to change the
functionality (e.g. to build a map of cell width that varies with
space).

The `IcosahedralMeshStep` class creates a subdivided icosahedral
mesh defined either by its approximate cell width or the number
of subdivisions.

A function `add_quasi_uniform_spherical_mesh_step` can be used to
add either of these last 2 types of steps to a test case,
depending on the config options.  This would allow a user to
decide whether to work with an icosahedral or a less uniform
"jigsaw" version of a given mesh.