Fix GIS test for current version of Albany
The greenland albany_input.yaml file was out of date and resulting in
garbage results.  Albany had changed units on flowA.  But we also had
added a realistic temperature field to the input file, so I've removed
the Problem section entirely so we use temp-dependent flow rate
calculation.  I added betaSolve to the output for convenience.