Update MALI submodule & adjust EISMINT2 test case
This merge updates the MALI submodule to bring in the newly merged PR
https://github.com/MALI-Dev/E3SM/pull/26
This change necessitated a namelist change to the EISMINT2 tests to
maintain the existing (correct) functionality for those tests.

Note that the MALI update changes answers for the greenland tests, but
in that case we want to switch to the new functionality (ignoring
positive SMB in ice-free areas).