Don't remove `esmf` from conda environment
This update to the conda environment causes problems if it
happens in the middle of both building code with the build
matrix and running with the conda environment.  It was always
intended as a "safety" measure to ensure that ESMF comes from
spack when we want that.  There's no indication that isn't
happening anyway.