Set up 3km versions of Humboldt tests
This commit adds a 3km version of the Humboldt mesh for the calving
tests.  It introduces a mesh_type that can either be 1km or 3km.  It
makes uses of a streams template to set the appropriate input files.  It
also updates the tests in the calving integration suite to use the 3km
version, but leaves the 1km tests in the list of available tests.