Adjust FO decomp test tolerances to get tests to pass
This changes the comparison tolerances for the decomposition FO tests to
that required to account for the non-BFB decomp behavior of the FO
solver.  i.e. the tolerances are set to those required for a run with no
calving to pass.

I also updated the eigencalving parameter to be smaller to keep the
uncalved error from killing the run.  The previous value was small
enough for running without velocity solver, but was still large once the
FO solver was turned on.

With these changes, all tests in the FO calving suite pass, with the
exception of
landice_humboldt_mesh-3km_decomposition_test_velo-fo_calving-ismip6_retreat
which is a known problem to be addressed in MALI at a later date.