Make tests run a bit faster
This commit makes tests run faster in two ways:
1. Restart tests only integrate one year forward after the restart.
There is no need to run a long time after the restart to ensure a
BFB-restart.  There is an implicit assumption here that the
decomposition tests will also be run and serve as a longer integrating
smoke test in addition to a decomp test.
2. For the 'full physics' configurations, increase the time step from 4
months to 6 months on the 3km mesh so they run a little faster.  This is
to allow them to run in the full_integration suite without adding too
much additional time.