change start date from July 1 to Jan 1
The ISMIP6 forcings are defined on July 1 of each year, and
at one point, there was an issue with reading forcings requiring that
the start time match the forcing time on init.  That problem was fixed
in https://github.com/MALI-Dev/E3SM/pull/10 so we no longer need to
start in the middle of the year to use the ISMIP6 forcing.  To avoid
confusion, I'm changing all tests to starrt on 2007-01-01.

Also change all stream reference times to 2007-01-01.  This is not
necessary, because the default reference time of 0000-01-01 will yield
the same results.  But I'm changing it for explicitness.  A minor
conisderation is choosing a reference_time closer to the model clock
time should provide a very small improvement to the cost of timekeeping
operations related to i/o.