Add support for a matrix of compilers/MPIs
When you configure the compass environment, you can pass a list
of compilers and a corresponding list of MPI libraries, or you
can pass "all" for one or both of these.  If you leave either
blank (or both), the default for the machine is used.  If you
pass "all" for either or both, only supported compiler/MPI pairs
are used.