Update Kangerlussuaq.geojson
Use a slightly larger region defined in the geojson file to avoid
the glacier reaching the edge of the mesh. This may still need to be
extended further. However, testing does not indicate a large difference
with the previous version, and this adds >1000 cells in the 1 to 10km mesh
with possibly a minimal gain, so perhaps reverting to the previous version
will be desireable.