Make dt_convergence_test test suite
This is a configuration for investigating convergence of calving
behavior with timestep for calving.  Rather than duplicating the
test in a different test groups, I've instead implemented the test as a
test group, and it can be applied to three different meshes: MISMIP+,
Humboldt 3km, and Thwaites 4km.  It can also be applied to three
different calving laws: specified calving velocity, von Mises stress,
and eigencalving.  It can be applied using either a data velocity field
or the FO solver.  This implementation replaces the previous commits
that applied this same concept to just the MISMIP+ geometry.

In each test, a series of runs are conducted that repeat the same
simulation using different values of
config_adaptive_timestep_calvingCFL_fraction.  Then summary plots are
generated comparing the behavior for the different fraction values.
The calving law parameters are modified for the different configurations
as needed to achieve steady retreat at a rate that ensures the calving
CFL condition is more restrictive than the advective CFL, even for large
values of config_adaptive_timestep_calvingCFL_fraction.

The configurations with no velocity solver typically take a few minutes
each.  The ones with FO velocity take much longer, in some cases a few
hours, even after setting those configurations to run for a shorter
duration.  As coded up now, each simulation with different fractions are
run in serial.  Setting this up to run them in parallel could be a
strategy for getting faster results, if desired.

Another limitation is that as coded up now, the analysis/plotting script
is run in the validation step and therefore cannot be run manually.
Therefore it will only run if all of the individual runs run to
completion successfully.  This could be improved by moving this
operation to a separate step.