Calculate window_size using config options
By default, calculate window_size for calculating distance to
grounding line and ice edge from the config options high_dist
. There is also an option to manually set window_size if desired,
but it must not be smaller than high_dist. If it is,
a warning is printed to the log file and window_size is calculated instead
of using the manually set value.