Add a counter to the name of the temporary files
For an unknown reason, ncremapping of gamma0 values from the source files yields wrong gamma0 values
in the processed output files when the name of the temporary files are kept as
"combined.nc" and "remapped.nc" even when the temporary files are removed at the end of each
file processing, requiring the temporary file names to be unique.
This commit adds a counter to the temporary file created inside the loop.