Remove the config section for the ocean_basal testcase
Previously, the ocean basal testcase had a config option for the remapping method.
However, as we always want to use the "nearest neighborhood" method for this testcase,
we decide to hard code the method in the main code "process_basal_melt.py" and
remove the config option from the config file.