Add database_component parameter for input files
This may be useful when accessing files for a different core
(e.g. seaice files within a test case in the ocean core).

This merge also changes the way that subdirectories for the
local database are specified.  Instead of speacifying a database_root
for each core, a single root is used and the database_subdirs
for each core are hard-coded in database_subdirs.json