Support running files_for_e3sm with existing mesh
This merge is a major reorganization of files_for_e3sm to have
a superclass for each step to handle common tasks such as linking
to a restart file, detecting whether ice-shelf cavities are in the
mesh, and getting attributes of the mesh such as the short name
and the creation date.

The DiagnosticsFiles step has been split into 2 steps since making
mapping files uses MPI parallelism whereas making region masks uses
thread parallelism and these need to be handled separately.

The MakeDiagnosticsFiles test case has been removed in favor of
this more complete solution.