Start creating Thwaites UQ study framework
This commit starts putting together a framework for setting up and
running a basic Thwaites UQ study.  It creates a set number of ensemble
members with parameter values randomly selected from pre-set ranges.
Currently, only the von Mises stress threshold and speed limit are
included as parameters.  It currently sets up the runs, but I have not
yet been able to successfully run them due to an Albany runtime library
issue.

A number of shortcuts have been made that may need to be generalized in
the future:
* input file locations are hard-coded
* numTasks is hard coded for Cori
* we may want parameter vectors to come from a file instead of being
  generated within the test case
* it's unclear if it's possible to create job scripts for each run.
  currently, compass run would only be able to run these 20 tests in
  serial