Switch the spd lim param to basal friction exponent
This includes:
* updating parameter vector
* modifying muFriction for updated exponent.  This makes use of a new general
extrapolate module added at the shared landice level.  It also switches to
first copying the input file to the run work_dir so that it can be modified
in place.  The input file path is now specified in the config file, though we
may want to move it to the database.
* modifying exponent value in yaml file.  This required directly copying the
yaml file instead of using 'add_input_file' so that the yaml contents could
be modified within the setup method.  It also makes use of a yaml parser to
modify an entry.  The yaml parser changes the ordering, which might not be desired.
* Commenting a hanging level of the yaml file, as the yaml parser writes its
value out as Null, which Albany does not like.

Note: we may add the calving speed limit back in later, but for now we want to
restrict ourselves to two parameters.