Fix major bug in vM calving
vM calving requires flow A to be calculated in MPAS in addition to using
the Albany calculation!  This was an oversight in my initial setup and
led to vM calving rates being way too low.  This fixes that.

Also raise the calving speed limit to 30 km/yr by default so the speed
limit is not likely to interact unless we lower it as part of parameter
sensitivity tests.