Move parameter value calculation to test case configure
The choice of what parameters to use and their desired ranges are now
specified in the test case configure and should be modified there if
users desire.  This is a cleaner approach. Also generalize parameter
ranges using a 2d unit Sobol sequence with 100 entries that can be used
for any variables and for any number of samples below 100.  It also
makes it easier to make parameter ranges config options if desired in
the future.