Major reorganzation of ensemble framework
Move the ensemble test case to its own group with thwaites as a test
case (instead of in the thwaites test group with uq_ensemble as the test
case).  There is little about this framework that is specific to
Thwaites Glacier, and it will be more flexible and less confusing to
have a dedicated ensemble test group.  This commit moves/renames the
files and modules and does a little cleanup.  A bit more cleanup is
necessary to make the new ensemble test group not specific to Thwaites.
I confirmed the test case runs and can set up an ensemble same as
before.