Use compass for job script and running model
This commit adjusts the previous design philosophy to make use of
compass functionality for:
1. creating graph files
2. creating job scripts
3. running the model

A separate job script is created for each step (ismip6 experiment).  The
number of procs to be used is defined by ntasks in the config.  The experiment
is run by submitting the job script in each experiment directory.  There is
no ‘compass run’ functionality at the test case level.