Change algorithm for constraining resources
Previously, nothing was preventing `cpus_per_task` from being
more than the number of cores on a node.  This is inconvenient
because unsafe or non-performant behavior will occur if more
python threads are used than cores on a node.

This merge changes the way resources are constrained.  First,
`cpus_per_task` is constrained to be less than the lower of
the number of cpus on a node or the number of total cores available.
Then, we check to see if `cpus_per_task` is smaller than the
minimum allowed for the given step.  Next, we allow all tasks
to have the same `cpus_per_task` and constrian the number of
tasks not to exceed the total available resources. Finally, we
check to make sure the number of tasks is not below the allowed
minimum for the step.

There are 4 typical resource requirements for steps:
* serial (1 task and 1 cpu per task)
* threading or multiprocessing on a node (1 task, multiple cpus per
  task)
* MPI jobs without threading (multiple tasks, 1 cpu per task)
* MPI jobs with threading (multiple tasks, a few cpus per task)

This algorithm should handle all of these without difficulty.