Switch to 64 vertical levels for 3 meshes
For the EC30to60, SO12to60 and WC14 meshes, switch from 60 to
64 vertical levels.  The surface and deep-ocean resolutions are
the same as before (10 and 250 m, respecitively) as is the
bottom depth (5500 m).

For the ARRM10to60 mesh, we stick with 80 vertical levels.  These
have thicknesses between 2 and 150 m, and a bottom depth of
5500 m.  These values are quite similar to but not identical to
the E3SM v1 80-layer mesh, see:

https://github.com/MPAS-Dev/compass/blob/main/compass/ocean/vertical/80layerE3SMv1.json