Move pio stride to cfg file
This is to avoid having it hard-coded.

This commit also moves the mesh resolution to the test case config
routine, because it was no longer used for anything in the
set_up_experiment module.  I added the mesh resolution to the subdir
where the test is set up to follow ISMIP6 naming convention.