Skip cached steps when computing suite resources
This merge skips cached steps when figuring out the target and
minimum cores for a suite.  This is important because the cached
steps require no resources at all.

It also sets the number of resources to 1 core by default (rather
than zero) in case there are no uncached steps in a run.