Allow a mix of user-defined and dataset-defined bounds
Allow the user to define the bounds around a regional domain with
any combination of floats and 'None'. For example, a config file that
contained ``x_min = 'None'; x_max = 1.0e6; y_min = -1.0e6; y_max =
'None'`` would use you left and top boundaries of the gridded dataset,
but also the user-defined values for the right and bottom boundaries.