Fix bug in changing permissions on downloaded files
The code for changing permissions was geing given an incorrect
directory and then was failing silently.  This merge fixes the
path and also removes the silent failure (an error will be raised
if the status of the download directory cannot be determined).

This merge also adds the database root and the root for each
affected core to the list of directories to chmod/chown anytime
any files get downloaded.