Fix substitution of `levels` at runtime
We need a different placeholder in config options for situations
where we need to replace them (not using extended interpolation
in the config file) at runtime.  Currently, this only applies to
`levels`, the number of vertical levels but it could apply to
`creation_date` and `bottom_depth` so we support these in the
same way.