Add thickness advection version of horizontal_advection test
This is identical to the horizontal_advection test of a passive tracer, but Gaussian
bump is applied to thickness field and tracer advection is not considered.  This
isolates the analysis to a single advection call; if there are issues with the
thickness advection, we presumably cannot expect tracer advection to converge properly.

It might have been possible to create an argument to toggle between
versions, but it was easier (and arguably less confusing) to just make a
copy.

This uses a Gaussian bump of 1000 m sitting on top of a 1000 m base
layer in a doubly periodic mesh.  This ensures a margin-less domain,
which would avoid the issue of advection degrading to 1st order at
margins.