New Mesh: SO12to60E3r2
Long name: SO12to60kmL64E3SMv3r2

This version of the Southern Ocean Regionally Refined Mesh (SORRM) has
the same distribution of resolution as in
[SOwISC12to60E3r1](https://github.com/MPAS-Dev/compass/pull/669):
* 12 km resolution around Antarctica
* 45 km resolution at southern mid-latitudes
* 30 km resolution at the equator and the north Atlantic
* 60 km resolution in the north Pacific
* 35 km resolution in the Arctic

It matches the EC30to60 mesh except in the Southern Ocean and north Atlantic.

This version of the mesh does not include the ice-shelf cavities around
Antarctica.