Fix a bug in how `landIceMask` is computed
Before this fix, cells that were more than half land according
to the remapped topography dataset but had been included in the
ocean (through critical passages) were mistakenly being
categorized as land ice simply because they were "not ocean".

With this fix, the `landIceMask` is computed based on where the
`landIceFracObserved` from the remapped topography dataset is
greater than 0.5.  A new function is needed to compute this mask
because this definition is not the same as the mask used to cull
the mesh when ice-shelf cavities are excluded from the ocean
domain.