Add data interpolation to AIS mesh generation test case
This commit Interpolates data from gridded datasets to an Antarctic mesh.
This takes care of the peculiarities of the current gridded compilation dataset
(antarctica_8km_2020_10_20.nc), as well as using conservative remapping directly
from the high-resolution BedMachineAntarctica and MeASUReS velocity datasets.
There is a fairly heavy degree of pre-processing done to get the BedMachine and
MeASUReS datasets ready to be used here. The pre-processing includes renaming
variables, setting reasonable _FillValue and missing_value attributes, extrapolating
fields to avoid interpolation ramps at ice margins, updating mask values, and
raising the bed topography at Lake Vostok to ensure a flat ice surface. Those
data files and processing scripts currently live here on Chicoma:
/usr/projects/climate/trhille/data
Eventually that pre-processing could be integrated into a new step in COMPASS,
or the processed data files could be added to the server on Anvil and downloaded
as needed.

The conservative remapping step using ESMF_RegridWeightGen requires multiple
nodes in order to not run out of memory.
This should probably be updated to be a separate step in the future in order to save on cost.

Co-authored-by: Trevor Hillebrand <trhille@lanl.gov>